requirements.txt
<<<<<<< HEAD
streamlit==1.32.0
pdfplumber==0.10.3
pandas==2.1.4
python-dateutil==2.8.2
pytesseract==0.3.10
opencv-python==4.8.1.78
pillow==10.1.0
numpy==1.26.2
openpyxl==3.1.2
pdf2image==1.16.3 
=======
streamlit==1.32.0
pdfplumber==0.10.3
pandas==2.1.4
python-dateutil==2.8.2
pytesseract==0.3.10
opencv-python==4.8.1.78
pillow==10.1.0
numpy==1.26.2
openpyxl==3.1.2
pdf2image==1.16.3 
>>>>>>> da9861285137d7690c9c82b14b5d0efeb26e646c
